"ads2_m12_a1_1: 2/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2 assignments/m12/ADS-2_Exam-2/Solution.java
@@ -45,7 +45,7 @@ public class Solution {
 			// Other wise print "No Path Found."
 			String[] places = scan.nextLine().split(" ");
 			Shortestpath sp = new Shortestpath(eg, Integer.parseInt(places[0]));
-			System.out.printf("%d\n", (float) sp.distTo(Integer.parseInt(places[1])));
+			System.out.printf("%d\n", (int) sp.distTo(Integer.parseInt(places[1])));
 			break;
 
 		case "ViaPaths":
